fix: fix bug open modal after login
